add PluginIssue trait
