+++ src/Phan/Plugin.php
@@ -19,6 +19,8 @@ use ast\Node;
  * of themselves.
  */
 abstract class Plugin implements ClassAnalyzer, FunctionAnalyzer, MethodAnalyzer {
+    use PluginIssue;
+
     /**
      * Do a first-pass analysis of a node before Phan
      * does its full analysis. This hook allows you to
@@ -71,62 +73,5 @@ abstract class Plugin implements ClassAnalyzer, FunctionAnalyzer, MethodAnalyzer
         Node $node,
         Node $parent_node = null
     );
-
-    /**
-     * Emit an issue if it is not suppressed
-     *
-     * @param CodeBase $code_base
-     * The code base in which the issue was found
-     *
-     * @param Context $context
-     * The context in which the issue was found
-     *
-     * @param string $issue_type
-     * A name for the type of issue such as 'PhanPluginMyIssue'
-     *
-     * @param string $issue_message
-     * The complete issue message to emit such as 'class with
-     * fqsen \NS\Name is broken in some fashion'.
-     *
-     * @param int $severity
-     * A value from the set {Issue::SEVERITY_LOW,
-     * Issue::SEVERITY_NORMAL, Issue::SEVERITY_HIGH}.
-     *
-     * @param int $remediation_difficulty
-     * A guess at how hard the issue will be to fix from the
-     * set {Issue:REMEDIATION_A, Issue:REMEDIATION_B, ...
-     * Issue::REMEDIATION_F} with F being the hardest.
-     */
-    public function emitIssue(
-        CodeBase $code_base,
-        Context $context,
-        string $issue_type,
-        string $issue_message,
-        int $severity = Issue::SEVERITY_NORMAL,
-        int $remediation_difficulty = Issue::REMEDIATION_B,
-        int $issue_type_id = Issue::TYPE_ID_UNKNOWN
-    ) {
-        $issue = new Issue(
-            $issue_type,
-            Issue::CATEGORY_PLUGIN,
-            $severity,
-            $issue_message,
-            $remediation_difficulty,
-            $issue_type_id
-        );
-
-        $issue_instance = new IssueInstance(
-            $issue,
-            $context->getFile(),
-            $context->getLineNumberStart(),
-            []
-        );
-
-        Issue::maybeEmitInstance(
-            $code_base,
-            $context,
-            $issue_instance
-        );
-    }
 
 }
